+++ packages/runtime/src/interpreter/interpreter.ts
@@ -136,14 +136,14 @@ export class Interpreter {
 	 * When `create` flag is false we will create new account and add it to context.
 	 * @param accountPk public key of account
 	 * @param line line number in TEAL file
-	 * @param create create flag
+	 * @param createFlag create flag
 	 * @param immutable allow to access foreign application account or not(false), default is true
 	 * https://developer.algorand.org/articles/introducing-algorand-virtual-machine-avm-09-release/
 	 */
 	private _getAccountFromAddr(
 		accountPk: Uint8Array,
 		line: number,
-		create: boolean,
+		createFlag: boolean,
 		immutable: boolean
 	): AccountStoreI {
 		const txAccounts = this.runtime.ctx.tx.apat; // tx.Accounts array
@@ -183,7 +183,7 @@ export class Interpreter {
 				) !== undefined)
 		) {
 			const address = encodeAddress(pkBuffer);
-			const account = create
+			const account = createFlag
 				? this.createAccountIfAbsent(address)
 				: this.runtime.ctx.state.accounts.get(address);
 
@@ -203,42 +203,61 @@ export class Interpreter {
 	 * When `create` flag is false we will create new account and add it to context.
 	 * @param accountRef index of account to fetch from account list
 	 * @param line line number
-	 * @param create create flag, default is true
+	 * @param createFlag create flag, default is true
 	 * @param immutable allow to access foreign application account or not(false), default is true
 	 * NOTE: index 0 represents txn sender account
 	 */
 	getAccount(
 		accountRef: StackElem,
 		line: number,
-		create = false,
+		createFlag = false,
 		immutable = true
 	): AccountStoreI {
 		let account: AccountStoreI | undefined;
 		let address: string;
-		if (typeof accountRef === "bigint") {
-			if (accountRef === 0n) {
-				address = encodeAddress(this.runtime.ctx.tx.snd);
-				account = this.runtime.ctx.state.accounts.get(address);
-			} else {
-				const accIndex = accountRef - 1n;
-				checkIndexBound(Number(accIndex), this.runtime.ctx.tx.apat ?? [], line);
-				let pkBuffer;
-				if (this.runtime.ctx.tx.apat) {
-					pkBuffer = this.runtime.ctx.tx.apat[Number(accIndex)];
-				} else {
-					throw new Error("pk Buffer not found");
-				}
-				address = encodeAddress(pkBuffer);
-				account = create
-					? this.createAccountIfAbsent(address)
-					: this.runtime.ctx.state.accounts.get(address);
-			}
+
+		if (accountRef === 0n) {
+			address = encodeAddress(this.runtime.ctx.tx.snd);
+			account = this.runtime.ctx.state.accounts.get(address);
+			return this.runtime.assertAccountDefined(address, account, line);
+		} else if (typeof accountRef === "bigint") {
+			return this._getAccountFromReference(accountRef, line, createFlag);
 		} else {
-			return this._getAccountFromAddr(accountRef, line, create, immutable);
+			return this._getAccountFromAddr(accountRef, line, createFlag, immutable);
 		}
+
+	}
+	
+	/**
+	 * Queries accesible accounts by smart contract. It can be either accounts defined
+	 * in accounts field or accounts associated with apps declared in foreingApps array
+	 * @param accountRef index of account or app ID
+	 * @param line line number
+	 * @param createFlag create flag
+	 * @retuns AccountStoreI object
+	 */
+	private _getAccountFromReference(accountRef: bigint, line: number, createFlag: boolean): AccountStoreI {
+		let address: string;
+		// check if reference exists in foreign apps
+		if (this.runtime.ctx.tx.apfa !== undefined &&
+			this.runtime.ctx.tx.apfa.includes(Number(accountRef))) {
+			address = getApplicationAddress(accountRef);
+		} else { // reference to accounts
+			const accIndex = accountRef - 1n;
+			checkIndexBound(Number(accIndex), this.runtime.ctx.tx.apat ?? [], line);
+			if (this.runtime.ctx.tx.apat) {
+				address = encodeAddress(this.runtime.ctx.tx.apat[Number(accIndex)]);
+			} else {
+				throw new Error("pk Buffer not found");
+			}
+		}
+		const account = createFlag
+			? this.createAccountIfAbsent(address)
+			: this.runtime.ctx.state.accounts.get(address);
 
 		return this.runtime.assertAccountDefined(address, account, line);
 	}
+
 
 	/**
 	 * Queries appIndex by app reference (offset to foreignApps array OR index directly)